Capstone_project/Breast_cancer_capstone_project/requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
tensorflow
kagglehub